feat: ask for delete files

+++ classes/ReposFiles.py
@@ -1,9 +1,7 @@
 import os
 import subprocess
-from pprint import pprint
-
+from rich import print
 from classes.CsvHandler import CsvHandler
-
 
 class ReposFiles:
     PUSH_FILE = os.path.join(os.path.expanduser('~'), 'Documents', 'push-repos.txt')
@@ -12,9 +10,13 @@ class ReposFiles:
         self.csvHandler = CsvHandler()
         self.exclude_dirs = self.csvHandler.getExcludeDirs()
         self.exclude_for_pull = self.csvHandler.getExcludeForPull()
-        self.deleteFiles()
-        self.reposeWriteToFile(self.PUSH_FILE, self.exclude_dirs)
-        self.reposeWriteToFile(self.PUT_FILE, self.exclude_for_pull + self.exclude_dirs)
+        delete_files = input("[green]Do you want to delete the files? (y/n): ")
+        if delete_files.lower() == 'y':
+            self.deleteFiles()
+            self.reposeWriteToFile(self.PUSH_FILE, self.exclude_dirs)
+            self.reposeWriteToFile(self.PUT_FILE, self.exclude_for_pull + self.exclude_dirs)
+        else:
+            print("[red]Files not deleted.")
 
     def deleteFiles(self):
         os.system(f"rm {self.PUSH_FILE} {self.PUT_FILE}")
